refactor: feedback applied
