+++ src/utils/index.ts
@@ -1,29 +1,3 @@
-import allList from '../assets/data/all-articles.json';
-import techList from '../assets/data/tech-articles.json';
-import designList from '../assets/data/design-articles.json';
-import { Article } from '../types/index';
-
-/**
- * 네트워크로 API 통신 모킹
- * */
-
-const dataMap: Record<string, Article[]> = {
-  '/': allList as Article[],
-  '/tech': techList,
-  '/design': designList,
-};
-
-async function delay(ms: number): Promise<void> {
-  return new Promise((resolve) => {
-    setTimeout(resolve, ms);
-  });
-}
-
-async function fetchList(path: string): Promise<Article[]> {
-  await delay(200);
-  return dataMap[path];
-}
-
 /**
  * 주어진 날짜 문자열을 "YYYY년 MM월 DD일" 형식으로 변환하는 함수
  * @param {string} dateString - 변환할 날짜 문자열 (예: "2023-10-18T09:00:00+09:00")
@@ -42,4 +16,4 @@ function formatDate(dateString: string) {
   return `${year}년 ${month}월 ${day}일`;
 }
 
-export { fetchList, formatDate };
+export { formatDate };
